+++ ui/components/app/assets/asset-list/cells/generic-asset-cell-layout.tsx
@@ -7,7 +7,7 @@ import {
   JustifyContent,
 } from '../../../../../helpers/constants/design-system';
 
-export type GenericAssetCellLayoutProps = {
+type GenericAssetCellLayoutProps = {
   onClick?: () => void;
   disableHover?: boolean;
   badge: ReactNode;
