+++ ui/components/app/assets/defi-list/cells/defi-protocol-cell.tsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react';
 import { useSelector } from 'react-redux';
-import { DeFiProtocolPosition } from '../defi-list';
 import GenericAssetCellLayout from '../../asset-list/cells/generic-asset-cell-layout';
 import { getPreferences } from '../../../../../selectors';
 
@@ -22,13 +21,17 @@ import {
   MetaMetricsEventName,
 } from '../../../../../../shared/constants/metametrics';
 import { MetaMetricsContext } from '../../../../../contexts/metametrics';
+import { DeFiProtocolPosition } from '../../types';
 
 type DeFiProtocolCellProps = {
   onClick: (chainId: string, protocolId: string) => void;
   position: DeFiProtocolPosition;
 };
 
-export function DefiProtocolCell({ onClick, position }: DeFiProtocolCellProps) {
+export default function DefiProtocolCell({
+  onClick,
+  position,
+}: DeFiProtocolCellProps) {
   const { privacyMode } = useSelector(getPreferences);
   const trackEvent = useContext(MetaMetricsContext);
