+++ ui/components/app/assets/defi-list/defi-tab.tsx
@@ -3,13 +3,11 @@ import { AssetListProps } from '../asset-list/asset-list';
 import AssetListControlBar from '../asset-list/asset-list-control-bar';
 import DefiList from './defi-list';
 
-const DeFiTab = ({ onClickAsset }: AssetListProps) => {
+export default function DeFiTab({ onClickAsset }: AssetListProps) {
   return (
     <>
       <AssetListControlBar showImportTokenButton={false} />
       <DefiList onClick={onClickAsset} />
     </>
   );
-};
-
-export default DeFiTab;
+}
